crawler/requirements.txt
bs4==0.0.1
requests==2.26.0
grpcio==1.41.0
